+++ release_notes.md
@@ -1,3 +1,9 @@
+## 07 - Fetching Articles
+- Fetches the articles data using a command
+- Display actual articles data
+- Pagination, maybe on a next article
+
+
 ## 06 - Home page
 - Split files: View, State, Types
 - Add packages: Fable.RemoteData Fable.Fetch Fable.Promise Thoth.Json